+++ src/betterproto/plugin/models.py
@@ -35,6 +35,34 @@ from dataclasses import dataclass, field
 from typing import Dict, Iterator, List, Optional, Set, Text, Type, Union
 
 from .. import Message, which_one_of
+
+import builtins
+import re
+import textwrap
+from dataclasses import dataclass, field
+from typing import Dict, Iterable, Iterator, List, Optional, Set, Type, Union
+
+import betterproto
+from betterproto import which_one_of
+from betterproto.casing import sanitize_name
+from betterproto.compile.importing import get_type_reference, parse_source_type_name
+from betterproto.compile.naming import (
+    pythonize_class_name,
+    pythonize_field_name,
+    pythonize_method_name,
+)
+from betterproto.lib.google.protobuf import (
+    DescriptorProto,
+    EnumDescriptorProto,
+    Field,
+    FieldDescriptorProto,
+    FieldDescriptorProtoLabel,
+    FieldDescriptorProtoType,
+    FileDescriptorProto,
+    MethodDescriptorProto,
+)
+from betterproto.lib.google.protobuf.compiler import CodeGeneratorRequest
+
 from ..casing import sanitize_name
 from ..compile.importing import get_type_reference, parse_source_type_name
 from ..compile.naming import (
@@ -54,7 +82,6 @@ from ..lib.google.protobuf import (
     ServiceDescriptorProto,
 )
 from ..lib.google.protobuf.compiler import CodeGeneratorRequest
-
 # Create a unique placeholder to deal with
 # https://stackoverflow.com/questions/51575931/class-inheritance-in-python-3-7-dataclasses
 PLACEHOLDER = object()
@@ -132,17 +159,13 @@ def get_comment(
                 sci_loc.leading_comments.strip().replace("\n", ""), width=79 - indent
             )
 
-            if path[-2] == 2 and path[-4] != 6:
-                # This is a field
-                return f"{pad}# " + f"\n{pad}# ".join(lines)
+            # This is a field, message, enum, service, or method
+            if len(lines) == 1 and len(lines[0]) < 79 - indent - 6:
+                lines[0] = lines[0].strip('"')
+                return f'{pad}"""{lines[0]}"""'
             else:
-                # This is a message, enum, service, or method
-                if len(lines) == 1 and len(lines[0]) < 79 - indent - 6:
-                    lines[0] = lines[0].strip('"')
-                    return f'{pad}"""{lines[0]}"""'
-                else:
-                    joined = f"\n{pad}".join(lines)
-                    return f'{pad}"""\n{pad}{joined}\n{pad}"""'
+                joined = f"\n{pad}".join(lines)
+                return f'{pad}"""\n{pad}{joined}\n{pad}"""'
 
     return ""
 
@@ -154,6 +177,8 @@ class ProtoContentBase:
     path: List[int]
     comment_indent: int = 4
     parent: Union["Message", "OutputTemplate"]
+
+    __dataclass_fields__: Dict[str, object]
 
     def __post_init__(self) -> None:
         """Checks that no fake default fields were left as placeholders."""
@@ -220,6 +245,7 @@ class OutputTemplate:
     imports: Set[str] = field(default_factory=set)
     datetime_imports: Set[str] = field(default_factory=set)
     typing_imports: Set[str] = field(default_factory=set)
+    builtins_import: bool = False
     messages: List["MessageCompiler"] = field(default_factory=list)
     enums: List["EnumDefinitionCompiler"] = field(default_factory=list)
     services: List["ServiceCompiler"] = field(default_factory=list)
@@ -251,6 +277,8 @@ class OutputTemplate:
         imports = set()
         if any(x for x in self.messages if any(x.deprecated_fields)):
             imports.add("warnings")
+        if self.builtins_import:
+            imports.add("builtins")
         return imports
 
 
@@ -266,6 +294,7 @@ class MessageCompiler(ProtoContentBase):
         default_factory=list
     )
     deprecated: bool = field(default=False, init=False)
+    builtins_types: Set[str] = field(default_factory=set)
 
     def __post_init__(self) -> None:
         # Add message to output file
@@ -359,6 +388,8 @@ class FieldCompiler(MessageCompiler):
         betterproto_field_type = (
             f"betterproto.{self.field_type}_field({self.proto_obj.number}{field_args})"
         )
+        if self.py_name in dir(builtins):
+            self.parent.builtins_types.add(self.py_name)
         return f"{name}{annotations} = {betterproto_field_type}"
 
     @property
@@ -366,6 +397,8 @@ class FieldCompiler(MessageCompiler):
         args = []
         if self.field_wraps:
             args.append(f"wraps={self.field_wraps}")
+        if self.optional:
+            args.append(f"optional=True")
         return args
 
     @property
@@ -391,9 +424,16 @@ class FieldCompiler(MessageCompiler):
             imports.add("Dict")
         return imports
 
+    @property
+    def use_builtins(self) -> bool:
+        return self.py_type in self.parent.builtins_types or (
+            self.py_type == self.py_name and self.py_name in dir(builtins)
+        )
+
     def add_imports_to(self, output_file: OutputTemplate) -> None:
         output_file.datetime_imports.update(self.datetime_imports)
         output_file.typing_imports.update(self.typing_imports)
+        output_file.builtins_import = output_file.builtins_import or self.use_builtins
 
     @property
     def field_wraps(self) -> Optional[str]:
@@ -415,6 +455,10 @@ class FieldCompiler(MessageCompiler):
         )
 
     @property
+    def optional(self) -> bool:
+        return self.proto_obj.proto3_optional
+
+    @property
     def mutable(self) -> bool:
         """True if the field is a mutable type, otherwise False."""
         return self.annotation.startswith(("List[", "Dict["))
@@ -429,10 +473,12 @@ class FieldCompiler(MessageCompiler):
         )
 
     @property
-    def default_value_string(self) -> Union[Text, None, float, int]:
+    def default_value_string(self) -> str:
         """Python representation of the default proto value."""
         if self.repeated:
             return "[]"
+        if self.optional:
+            return "None"
         if self.py_type == "int":
             return "0"
         if self.py_type == "float":
@@ -443,6 +489,14 @@ class FieldCompiler(MessageCompiler):
             return '""'
         elif self.py_type == "bytes":
             return 'b""'
+        elif self.field_type == "enum":
+            enum_proto_obj_name = self.proto_obj.type_name.split(".").pop()
+            enum = next(
+                e
+                for e in self.output_file.enums
+                if e.proto_obj.name == enum_proto_obj_name
+            )
+            return enum.default_value_string
         else:
             # Message type
             return "None"
@@ -487,9 +541,14 @@ class FieldCompiler(MessageCompiler):
 
     @property
     def annotation(self) -> str:
+        py_type = self.py_type
+        if self.use_builtins:
+            py_type = f"builtins.{py_type}"
         if self.repeated:
-            return f"List[{self.py_type}]"
-        return self.py_type
+            return f"List[{py_type}]"
+        if self.optional:
+            return f"Optional[{py_type}]"
+        return py_type
 
 
 @dataclass
@@ -624,53 +683,20 @@ class ServiceMethodCompiler(ProtoContentBase):
         self.parent.methods.append(self)
 
         # Check for imports
-        if self.py_input_message:
-            for f in self.py_input_message.fields:
-                f.add_imports_to(self.output_file)
         if "Optional" in self.py_output_message_type:
             self.output_file.typing_imports.add("Optional")
-        self.mutable_default_args  # ensure this is called before rendering
 
         # Check for Async imports
         if self.client_streaming:
             self.output_file.typing_imports.add("AsyncIterable")
             self.output_file.typing_imports.add("Iterable")
             self.output_file.typing_imports.add("Union")
-        if self.server_streaming:
+
+        # Required by both client and server
+        if self.client_streaming or self.server_streaming:
             self.output_file.typing_imports.add("AsyncIterator")
 
         super().__post_init__()  # check for unset fields
-
-    @property
-    def mutable_default_args(self) -> Dict[str, str]:
-        """Handle mutable default arguments.
-
-        Returns a list of tuples containing the name and default value
-        for arguments to this message who's default value is mutable.
-        The defaults are swapped out for None and replaced back inside
-        the method's body.
-        Reference:
-        https://docs.python-guide.org/writing/gotchas/#mutable-default-arguments
-
-        Returns
-        -------
-        Dict[str, str]
-            Name and actual default value (as a string)
-            for each argument with mutable default values.
-        """
-        mutable_default_args = {}
-
-        if self.py_input_message:
-            for f in self.py_input_message.fields:
-                if (
-                    not self.client_streaming
-                    and f.default_value_string != "None"
-                    and f.mutable
-                ):
-                    mutable_default_args[f.py_name] = f.default_value_string
-                    self.output_file.typing_imports.add("Optional")
-
-        return mutable_default_args
 
     @property
     def py_name(self) -> str:
@@ -728,6 +754,17 @@ class ServiceMethodCompiler(ProtoContentBase):
             imports=self.output_file.imports,
             source_type=self.proto_obj.input_type,
         ).strip('"')
+
+    @property
+    def py_input_message_param(self) -> str:
+        """Param name corresponding to py_input_message_type.
+
+        Returns
+        -------
+        str
+            Param name corresponding to py_input_message_type.
+        """
+        return pythonize_field_name(self.py_input_message_type)
 
     @property
     def py_output_message_type(self) -> str: