+++ tests/util.py
@@ -1,8 +1,12 @@
 import importlib
+import asyncio
+from dataclasses import dataclass
 import os
 import pathlib
 from pathlib import Path
+import sys
 from types import ModuleType
+from typing import Callable, Dict, Generator, List, Optional, Tuple, Union
 from typing import Callable, Generator, List, Optional
 
 os.environ["PROTOCOL_BUFFERS_PYTHON_IMPLEMENTATION"] = "python"
@@ -19,10 +23,45 @@ def get_directories(path: Path) -> Generator[str, None, None]:
 
 
 def get_test_case_json_data(test_case_name: str, *json_file_names: str) -> List[str]:
+async def protoc(
+    path: Union[str, Path], output_dir: Union[str, Path], reference: bool = False
+):
+    path: Path = Path(path).resolve()
+    output_dir: Path = Path(output_dir).resolve()
+    python_out_option: str = "python_betterproto_out" if not reference else "python_out"
+    command = [
+        sys.executable,
+        "-m",
+        "grpc.tools.protoc",
+        f"--proto_path={path.as_posix()}",
+        f"--{python_out_option}={output_dir.as_posix()}",
+        *[p.as_posix() for p in path.glob("*.proto")],
+    ]
+    proc = await asyncio.create_subprocess_exec(
+        *command, stdout=asyncio.subprocess.PIPE, stderr=asyncio.subprocess.PIPE
+    )
+    stdout, stderr = await proc.communicate()
+    return stdout, stderr, proc.returncode
+
+
+@dataclass
+class TestCaseJsonFile:
+    json: str
+    test_name: str
+    file_name: str
+
+    def belongs_to(self, non_symmetrical_json: Dict[str, Tuple[str, ...]]):
+        return self.file_name in non_symmetrical_json.get(self.test_name, tuple())
+
+
+def get_test_case_json_data(
+    test_case_name: str, *json_file_names: str
+) -> List[TestCaseJsonFile]:
     """
     :return:
-        A list of all files found in "inputs_path/test_case_name" with names matching
-        f"{test_case_name}.json" or f"{test_case_name}_*.json", OR given by json_file_names
+        A list of all files found in "{inputs_path}/test_case_name" with names matching
+        f"{test_case_name}.json" or f"{test_case_name}_*.json", OR given by
+        json_file_names
     """
     test_case_dir = inputs_path / test_case_name
     possible_file_paths = [
@@ -35,7 +74,11 @@ def get_test_case_json_data(test_case_name: str, *json_file_names: str) -> List[
     for test_data_file_path in possible_file_paths:
         if not test_data_file_path.exists():
             continue
-        result.append(test_data_file_path.read_text())
+        result.append(
+            TestCaseJsonFile(
+                test_data_file_path.read_text(), test_case_name, test_data_file_path.name.split(".")[0]
+            )
+        )
 
     return result
 
@@ -56,7 +99,7 @@ def find_module(
     if predicate(module):
         return module
 
-    module_path = pathlib.Path(*module.__path__)
+    module_path = Path(*module.__path__)
 
     for sub in [sub.parent for sub in module_path.glob("**/__init__.py")]:
         if sub == module_path: